+++ source/lib/lua/script_host.h
@@ -141,7 +141,7 @@ private:
    // Memory profiling
    bool                 enable_profile_memory_;
    int                  _gc_setting;
-   typedef std::unordered_map<void*, int>          Allocations;
+   typedef std::unordered_map<void*, size_t>       Allocations;
    std::unordered_map<void *, std::string>         alloc_backmap;
    std::unordered_map<std::string, Allocations>    alloc_map;
